% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{blast_help}
\alias{blast_help}
\title{Print the help function}
\usage{
blast_help(tool = "blastx")
}
\arguments{
\item{tool}{blast tools: blastn, blastp, blastx}
}
\description{
Print the help function
}
\examples{
blast_help()
}
\author{
Kai Guo
}
