% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.find.tools}
\alias{.find.tools}
\title{find excute blast or diamond}
\usage{
.find.tools(tool = "blastp")
}
\arguments{
\item{tool}{tools: blastp, blastx, blastn}
}
\description{
find excute blast or diamond
}
\author{
Kai Guo
}
