% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedb.R
\name{.get.spe.code}
\alias{.get.spe.code}
\title{find the species code for kegg}
\usage{
.get.spe.code(species = "hsa")
}
\arguments{
\item{species}{character, either the kegg code, scientific name or the
common name of the target species. Default species="hsa", it is
equivalent to use either "Homo sapiens" (scientific name) or
"human" (common name).}
}
\description{
find the species code for kegg
}
