% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{.makeblastdb}
\alias{.makeblastdb}
\title{makeblastdb command}
\usage{
.makeblastdb(filepath, dbtype = "prot", args = "")
}
\arguments{
\item{filepath}{path of fasta file}

\item{dbtype}{character "nucl" or "prot"}

\item{args}{other arguments for makeblastdb}
}
\description{
makeblastdb command
}
\author{
Kai Guo
}
