% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{.makediamonddb}
\alias{.makediamonddb}
\title{diamond makedb command}
\usage{
.makediamonddb(filepath, args = "")
}
\arguments{
\item{filepath}{path of fasta file}

\item{args}{other arguments for diamond makedb}
}
\description{
diamond makedb command
}
\author{
Kai Guo
}
