% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedb.R
\name{.vsplit}
\alias{.vsplit}
\title{split vector into chunk with provide length}
\usage{
.vsplit(v, n)
}
\arguments{
\item{v}{vector}

\item{n}{number of chunks}
}
\description{
split vector into chunk with provide length
}
