% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedb.R
\name{get.kegg.code}
\alias{get.kegg.code}
\title{extract kegg code
modified from pathview kegg.species.code}
\usage{
get.kegg.code(species = "hsa")
}
\description{
extract kegg code
modified from pathview kegg.species.code
}
