% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedb.R
\name{get.kegg.entry}
\alias{get.kegg.entry}
\title{collect all the pathway name of species from KEGG}
\usage{
get.kegg.entry(specode)
}
\arguments{
\item{specode}{KEGG species code}
}
\value{
a data frame with pathways information
}
\description{
collect all the pathway name of species from KEGG
}
\author{
Kai Guo
}
