% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedb.R
\name{keggGetm}
\alias{keggGetm}
\title{Function modified from KEGGREST}
\usage{
keggGetm(dbentries, option = c("aaseq", "ntseq", "mol", "kcf", "image",
  "kgml"))
}
\arguments{
\item{dbentries}{One or more (up to a maximum of 10) KEGG identifiers.}

\item{option}{Optional. Option governing the format of the output.
aaseq is an amino acid sequence, ntseq is a nucleotide sequence. image
returns an object which can be written to a PNG file, kgml returns a
KGML document.}
}
\description{
Function modified from KEGGREST
}
