% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{korg}
\alias{korg}
\title{korg}
\format{A data frame including including taxonomy IDs, KEGG code,
scientific name, common name, corresponding gene ID types, and
gene annotation package names in Bioconductor}
\usage{
korg
}
\description{
korg is the dataset from pathview
}
\examples{
korg
}
\keyword{datasets}
