% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{makeblastdb}
\alias{makeblastdb}
\title{make blast database}
\usage{
makeblastdb(db, seqcode = NULL, dbtype = "prot", args = "",
  path = NULL, runblast = TRUE)
}
\arguments{
\item{db}{KEGG db or sequences}

\item{seqcode}{speceis name or KEGG code (without spaces)}

\item{dbtype}{character "nucl" or "prot"}

\item{args}{other arguments for makeblastdb}

\item{path}{working directory}

\item{runblast}{use blast+ or diamond}
}
\description{
make blast database
}
\author{
Kai Guo
}
