% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedb.R
\name{preparedb}
\alias{preparedb}
\title{Prepare pathway sequences database from KEGG}
\usage{
preparedb(species = "human", seqtype = "AA", savedb = TRUE,
  path = NULL)
}
\arguments{
\item{species}{species name for KEGG parse}

\item{seqtype}{sequence type: AA or DNA (default: AA)}

\item{savedb}{save the kegg database for future using.}

\item{path}{path for saving kegg database (savedb = TRUE,default="./")}
}
\value{
list with sequences and annotation for pathway
}
\description{
Prepare pathway sequences database from KEGG
}
\author{
Kai Guo
}
