% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{seq2fun}
\alias{seq2fun}
\title{Blast search against kegg database}
\usage{
seq2fun(query, dbfile, anno, type = "blastp", evalue = 1e-10,
  num_threads = 1)
}
\arguments{
\item{query}{query sequences}

\item{dbfile}{database}

\item{anno}{kegg annotation file}

\item{type}{blastx, blastn or tblast}

\item{evalue}{e-value}

\item{num_threads}{num_threads}
}
\description{
seq2fun provide function to annotate sequences
}
\author{
Kai Guo
}
