% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedb.R
\name{get.kegg.code}
\alias{get.kegg.code}
\title{Extract kegg code}
\usage{
get.kegg.code(species = "hsa")
}
\arguments{
\item{species}{species user provide}
}
\description{
get.kegg.code are modified from pathview kegg.species.code
}
