% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{makeblastdb}
\alias{makeblastdb}
\title{make blast database}
\usage{
makeblastdb(file, dbtype = "prot", args = "")
}
\arguments{
\item{file}{sequence file}

\item{dbtype}{character "nucl" or "prot"}

\item{args}{other arguments for makeblastdb}
}
\description{
make blast database
}
\author{
Kai Guo
}
