% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{seq2fun}
\alias{seq2fun}
\title{Blast search against kegg database}
\usage{
seq2fun(query, db, type = "blastp", evalue = 1e-10, num_threads = 1,
  path = NULL, bidirectional = FALSE, runblast = TRUE)
}
\arguments{
\item{query}{query sequences}

\item{db}{database}

\item{type}{blastx, blastn or tblast}

\item{evalue}{e-value}

\item{num_threads}{num_threads}

\item{path}{working directory}

\item{bidirectional}{use bidirectional blast}

\item{runblast}{use blast+ or diamond}
}
\description{
seq2fun provide function to annotate sequences
}
\author{
Kai Guo
}
